refactor(storage): move the footprint models into the model layer
check_ui_service_boundary failed: the diagnostic widget imported
package:openvine/services/ for StorageFootprint and friends, which the
UI layer may not do. The models are data, not the service that produces
them, so they move to lib/models/storage_footprint.dart and the widget
reads them from there.
